BPM and Training Dummy
We currently have a working BPM to shoot mechanic. Find it under the SinglePlayer scene, or as BPM manager.

Changeable error windows for hitting on the correct time.

If you play the SinglePlayer scene it will play an 80 BPM metronome regardless of what you input into the BPM Manager because that is what I put in the file. The shooting buff will still call from the BPM manager to see if it does double damage or not.